+++ update/command.go
@@ -6,9 +6,9 @@ import (
 	"github.com/nothub/hashutils/chksum"
 	"github.com/nothub/hashutils/encoding"
 	"github.com/nothub/mrpack-install/files"
-	"github.com/nothub/mrpack-install/http/download"
 	"github.com/nothub/mrpack-install/modrinth/mrpack"
 	"github.com/nothub/mrpack-install/update/backup"
+	"github.com/nothub/mrpack-install/web/download"
 	"log"
 	"os"
 	"path"
