+++ web/client.go
@@ -1,4 +1,4 @@
-package http
+package web
 
 import (
 	"fmt"
